add useTerminalState composable
